Relicence to MPL
This is intended to be retroactive, any previous version of the code in this repository is also now available under the MPL as well as the AGPL, at your option.

From this commit, the code is solely available under the MPL, not the AGPL.

All copyright in this repository is owned by me (jamie@quigley.xyz), so re-licencing is possible.